--- src/ToastNotification.jsx
@@ -1,1 +1,0 @@
-
